Refactor. Initialize global and local variables by default.
