+++ src/regression/expr.rs
@@ -1,47 +1,59 @@
 // This file was generated by expr.py
 
 type Int = crate::types::Int;
-type Expr = crate::expr::Expr;
-type Context = crate::context::Memory;
+type Expr = crate::statement::Expr;
+type Context<'a> =  &'a [(&'static str, Int)];
 
 fn parse(input: &str) -> Expr {
     Expr::parse(input).unwrap()
 }
 
-fn eval(expr: Expr, memory: &Context) -> Int {
-    use crate::context::{EmptyInput, IgnoreOutput};
-    use crate::sm::{self, StackMachine};
-    use crate::typecheck::{self, ExecutionContext};
-    use crate::statement::{self, Statement};
+fn eval(expr: Expr, globals: &Context) -> Int {
+    use crate::memory::ScopedMemory;
+    use crate::io::{EmptyInput, IgnoreOutput};
+    use crate::sm;
+    use crate::syntax::Statement;
+    use crate::statement;
     use crate::types::Var;
 
-    let e = {
-        let program = typecheck::Program::empty();
-        let mut mem = memory.clone();
-        let mut input = EmptyInput;
-        let mut output = IgnoreOutput;
-        let mut context = ExecutionContext::new(&program, &mut mem, &mut input, &mut output);
-        // first evaluate expr in expression language
-        let e = expr.eval(&mut context).unwrap();
-        e
+    let new_memory = || {
+        let vars = globals.iter().map(|(name, _v)| name.to_string());
+        ScopedMemory::with_globals(vars)
     };
 
+    // first evaluate expression only
+    let e = {
+        let (_, program) = statement::Program::from_main(vec![
+            Statement::Return(Some(expr.clone()))
+            ], globals.iter().map(|(name, _value)| name.to_string())).unwrap();
+
+            let mut memory = new_memory();
+            for (name, value) in globals.iter() {
+                memory.store(&name.to_string(), *value);
+            }
+
+            let mut input = EmptyInput;
+            let mut output = IgnoreOutput;
+            let e = statement::run(&program, &mut memory, &mut input, &mut output).unwrap();
+            e.unwrap()
+        };
+
     let result_var = Var::from("RESULT");
-    // then in statements language
+    // then with statements
     let (s, program) = {
         let mut main = Vec::new();
-        for (name, value) in memory.globals().iter() {
+        for (name, value) in globals.iter() {
             let set_var = Statement::Assign(name.to_string(), Expr::Const(*value));
             main.push(set_var);
         }
         main.push(Statement::Assign(result_var.clone(), expr));
-        let program = statement::Program::from_main(main);
+        let (_, program) = statement::Program::from_main(main, std::iter::empty()).unwrap();
 
-        let (_, program) = crate::typecheck::Program::check(program).unwrap();
-        let mut memory = Context::new();
+        let mut memory = new_memory();
         let mut input = EmptyInput;
         let mut output = IgnoreOutput;
-        program.run(&mut memory, &mut input, &mut output).unwrap();
+        statement::run(&program, &mut memory, &mut input, &mut output).unwrap();
+
         let s = memory.load(&result_var).unwrap();
         (s, program)
     };
@@ -50,12 +62,12 @@ fn eval(expr: Expr, memory: &Context) -> Int {
     // then using state machine instructions
     let (sm, _program) = {
         let program = sm::compile(&program).unwrap();
-        let mut memory = Context::new();
+
+        let mut memory = new_memory();
         let mut input = EmptyInput;
         let mut output = IgnoreOutput;
-        let mut machine = StackMachine::new(&mut memory, &mut input, &mut output);
-        machine.run(&program).unwrap();
-        assert_eq!(machine.pop(), None); // stack should be empty
+        sm::run(&program, &mut memory, &mut input, &mut output).unwrap();
+
         let sm = memory.load(&result_var).unwrap();
         (sm, program)
     };
@@ -77,14 +89,8 @@ fn eval(expr: Expr, memory: &Context) -> Int {
     sm
 }
 
-fn make_context(variables: &[(&str, Int)]) -> Context {
-    let mut memory = Context::new();
+fn make_context<'a>(variables: &'a [(&'static str, Int)]) -> Context<'a> {
     variables
-        .iter()
-        .map(|(name, value)| (name.to_string(), *value))
-        .for_each(|(name, value)| memory.store(&name, value));
-
-    memory
 }
 
 #[test]
